\name{normtongue}
\alias{normtongue}
\title{Level Out Ultrasound Tongue Measurements Automatically}
\usage{
TODO...
}
\description{
Ultrasound tongue imaging is a non-invasive technique used in phonetics
experiments to obtain data on the position and shape of a speaker's tongue at
certain points in time.  A thin, stiff object such as a straightedge may be
entered into the speaker's mouth in order to provide an image of the occlusal
plane before the actual measurements, based on which the other data may be
oriented to the same plane.  This package provides a few basic, easy to use
functions to rotate all tongue image data into the plane indicated by the
straightedge.
}
\examples{
TODO...
}
